+++ mlxtend/regression/__init__.py
@@ -1,2 +1,4 @@
 from .lin_regplot import lin_regplot
-__all__ = ["lin_regplot"]
+from .linear_regression import LinearRegression
+
+__all__ = ["lin_regplot, LinearRegression"]